+++ msfs-multiplayer/src/simconnect.rs
@@ -65,7 +65,7 @@ struct UserAircraftData {
     velocity_body_z: f64,
     #[name = "SIM ON GROUND"]
     #[unit = "Bool"]
-    on_ground: i32,
+    on_ground: i64,
 }
 
 #[data_definition]
